Use an intermediate project to handle GitHub webhook
The GitHub plugin webhook handling is a bit shaky and it makes it
hard to run a job by selecting the branch / tag to run on it. Instead,
this create a jobs whose sole job is to parse the GitHub webhook
json payload and call the appropriate project with the good parameters.

Change-Id: Iaa55d3139b77f63dbb61c1edff9a1ee6e206f56d