Add build steps to actually run the build on Windows. /noideadog
For now, only for the Bazel step (I'll be happy if it works) and it's a miracle that it works at all: after checking all the documentation of Jenkins, I blindly checked if ${PLATFORM_NAME} works to condition running a step on the platform it's running on, and when it did, I experimented until I found out exactly what flavor of regular expression matching the "Run Condition" plugin does.

Good thing is, it now works.

Change-Id: I03a8108d9ec72adedc181ea386d17aaab7343c86