Use Jinja2 for templating engine for our configuration file
This enable to put more logic in the configuration file template, especially
it will enable us to have the shell scripts in a more human friendly format.

For now, just do a very basic replacement of the old broken template actions.

This was tested by comparing the result XML files, the only resulting change
in XML files where some whitespaces, some fixses of the plugin attribute (maybe
we could just remove it all together, this seem to be ignored by Jenkins), and
a &quote; becoming " (it was escaping without need for it to be).

Change-Id: I2e9052c211476931259f6f1c82fa6b89ecfbf10b