Disable Github trigger for bazel-tests
Until we set-up the automatic push, this tests is executed as
a downstream build of the Bazel job so it don't need to be executed
as a result of a Github push too.

Once we start pushing automatically our change to Github, we
will make the other job a nightly and have this one vet every push.

Change-Id: I278921ef7c7196f9ac9d22011933b4289adb0963