Add a gerrit verification job to ci.bazel.io
This work using the Jenkins Build Job Flow plugin
to query the Gerrit REST API and poll every minutes
for untested review.

To test a review, a reviewer mark it as verified (set
verified comment to +1) then jenkins should detect
it and build it and set the verified bit to +1 (test passes)
or -1 (test fails).

Fixes #11

Change-Id: I823bd2aa126a14ef97f1231bf008bd0369ec7fd6