Revert "Clean up the output base when the a Bazel Github test finishes."
This reverts commit aa26c97e3903eff045ed6ae1cefb057adba4cb33.

If the output base grows indefinitely this is a bug in Bazel and cleaning
it prevent it from being detected. Cleaning the output base also prevent
any caching, we don't have a distributed caching on ci.bazel.io and we
do not do selective cache. Let's not clean it.

Note that the original commit is cleaning the cache of all our other
project like rules_closure.

Change-Id: I65f0352405a39d950b44bade2be160bc160c82d9