Replace Tutorial job by a github job
This enables PR vetting but also ensures that we use the
same android setup that for Bazel itself.

As a drawback this is no longer used to prevent bad releases but we
can count on the release manager to not publish a bad release.

This was tested by deploying to ci-staging.bazel.io. However,
this still requires https://github.com/bazelbuild/examples/pull/16
to be merged first so cannot test the build itself. I did verify
that the configuration looked correct.

Change-Id: I38f443989e48634277f7cbdde66bf16dfd9cc0ee