Change CI to support a bootstrap bazel
As bootstrap version of bazel, we use the latest release of bazel.
For this to work, we have to make sure that bazel is installed on
every machine. In this way, CI will continue to work once we change
to building bazel from a bootstrap bazel instead of a simple
compile.sh from a checkout.

Change-Id: I247970139b8e85c3162e9fffda4cf4a4e11c3687