Add support for "local global libraries"
So Jenkins does support Global Pipeline Libraries, but it
knows only how to fetch it from a repository. There is a
filesystem_scm plugin for Jenkins but it does not interact
well with Global Pipeline Libraries. So the solution is
to use a local git repository.

Now here what this change does:
  - Create a git repository from lib that are passed through pkg_tar
    to the docker image in /opt/lib.
  - Since pkg_tar does not support setting owner, we change the permission
    by tweaking the entrypoint of the docker container before running the
    jenkins startup script.
  - GlobalLibraries configuration just look at a Git SCM on the jenkins
    master.

Change-Id: Ie61023c1f06fe5bdba63aa54cf83dc6a789aef70