Add BazelConfiguration groovy class
This is just a convenience class to read configuration from
hierarchical json descriptor and return a list of configuration
to execute. Each configuration contains a descriptor (key=value pair,
e.g. the node label to run on) and a set of configuration parameters.

This change also add a BUILD file and some test to the groovy library
so we need to change the way we ship the library inside the docker
container.

Change-Id: I5d6a2090b4360a0497b96e466df65ee52d19bbdb