Add a bazelJob pipeline step
This step execute bazel and will be used for running all the
job on CI using Jenkins pipeline.

This was tested on a local instance of ci, using a linux slave.
It mimick the way we are currently doing bazel testing, and should
work allright on Windows but was not fully tested yet.

Change-Id: If93663a3d7e9f2589fef8ec78df8330e8c3acc2f