Add two pipeline step: maybeDir and maybeSauce
maybeDir change directory (like the dir step) only if the
directory parameter is not empty. It will be used to change
directory if the workspace attribute is not empty.

maybeSauce launch sauce connect (like the sauce step) only if
the identifier parameter is not empty. Enabling sauce is just
a question of setting that parameter to the sauce credentials
identifier.

Change-Id: If5a6e29ff316fc21d0c805329a56f41ebffcba61