Add two steps that wrap the bazelJob step
bazelCiJob get bazel using fetchBazel, checkout the git repository and
execute the bazelJob step on a specific node type. It also set some
custom environment dedicated to our CI. Whereas bazelJob is a general
purpose step that we could offer as a library to everybody, bazelCiJob
is a step dedicated to our own CI.

bazelConfiguredCiJob take a JSON configuration string and expand it
into several bazelCiJob for the various configurations specified by
the configuration file.

Change-Id: Ic58f4177956ca9d9c1b46f156e33eebdefa0290a