Enable to use configuration files from the repository
The configuration files are currently checked into the CI system
repository. However, they really belongs to the repository, they
are the "blueprint" of what should be build and tested on the CI system.

The file can be checked in either at .ci/${name}.json or
at scripts/ci/${name}.json.

Change-Id: I2f0dedf7378ccb85605d0aa22d354e74b5a06ff7