Ensure our jenkins-slave startup script terminates
...by properly daemonizing the jenkins job. In this way, all startup
jobs terminate (after starting their respective daemons), and we no
longer depend on that job being started the last. In fact, on some
systems, other tasks are only started after after the termination of
the startup script---leading to the situation that jenkins works, but,
e.g., sshd is not running.

Note that still is not the proper way of starting a daemon; a proper
integration with the init system would be preferrable. As, however,
those differ between the various Unix-like systems, and we need to work
for all of them, this is a good trade-off.

Change-Id: I06653184ad09c547572c4744aaf74bd33ed1c9a5