Exclude -jdk7 artifacts when doing the release
For the next release we do not want to release JDK7 artifact, since
the release is going to deprecate it, simply exclude the artifact from
copy.

This change does not remove any other JDK7 path, we will remove it
once we passed one week with 0.5.2. Then we can allow introduction
of JDK8 feature and kill all the code path for JDK7 :)

Note: this is pretty hard to test so I will do a test release tomorrow
morning with this change in, avoiding the necessity for another canary
for 0.5.2.

Change-Id: Ib450383a17af30f85fc82f5b314f80d6c107b514