Report the Blue Ocean URL on Gerrit presubmit
This URL is easier to visualize with the new job and failure are
easier to browse. Although due to https://issues.jenkins-ci.org/browse/JENKINS-26522
still being worked on, we currently cannot mark a specific step as
"unstable", so all the step appear as unstable when a test fails. In which
case looking at the list of test result might be better, but that's one
click away from the default view in Blue Ocean.

Change-Id: I3d221276380b814cb0bd10fee08fffe16a18ca2c