runAll: do not set global status, instead report individually
Setting the global status completely hides where the cause was.
However, for our use, we the main information is which downstream
project had problems, the difference between failure (as in "fails
to build") and unstable ("builds, but tests fail") is only secondary.
So, use the pipelines in a more standard way let individual jobs
report their individual failures---while still cathching the errors
and continuing with the final reporting step.

Change-Id: I8db14e81bcb33d5f902b16257748f38d44b09d39