Also test bazel with remote execution
Bazel recently added support for remote execution of build actions.
This change enables Bazel’s CI system to run its builds against an
experimental remote execution and caching service. In the initial
verification stage, this will be done in an isolated build project, run
in parallel with our tested CI setup.

In order to keep the build steps similar for all builds, credentials
for the remote service are provided to all builds, whether or not remote
execution is enabled.

Change-Id: Ib7d986ab2004586fdb001c660ec2a779af8477f4