Do not vet PR for bazel-tests-remote
For some reason, on Github PR, 2 tests seemed to cancel
the second test if the first one fails. This gives us wrong
coverage if there is a flake.

Ideally we would fix the underlying issue but for now it
is more important to get bazel-tests vetting PR than to
get bazel-remote-tests.

Change-Id: I9061e69ed5753bd348b44fe956bfb82631cd2a05