Pass required Bazelisk env vars to test step (#489)
* Pass required Bazelisk env vars to test step.

Bazelisk replaces /usr/bin/bazel. As a consequence, any integration test that runs on a machine with Bazelisk and that invokes Bazel actually calls Bazelisk, but without access to the required environment variables (due to how "bazel test" works). Consequently, the invocation fails. This commit fixes the problem by passing the required environment variable ($HOME) to the test. For consistency, it also passes $USE_BAZEL_VERSION so that the test doesn't use a different version of Bazel.