Add --script flag to bazelci.py
This flag is meant for testing, since it allows the user to specify the URL of the bazelci.py file that should be retrieved by runner steps.
As a result, the URL can be easily specified in the pipeline configuration instead of having to replace the URL in the file itself.