Wrap an expected-to-fail call that in a script
In cmd.exe we cannot branch directly on exit code
like in Bash, we need to check ERRORLEVEL instead.

BuiltKite however checks ERRORLEVEL after every
call and fails-fast if the last call failed.

The "git merge-base" command may fail (we expect
that) so in order to branch on that, we wrap the
call in a .bat file.

See https://github.com/bazelbuild/continuous-integration/pull/752#issuecomment-512168511