Run "setup" tasks before entering the working directory
Tasks may now specify the optional "setup" field. It contains a list of commands which should be run *before* entering a custom working directory (as opposed to batch_commands and shell_commands that are run *inside* the working directory).

This mechanism is required by the Bazel federation and was already tested on Buildkite.