Fix a race condition when starting the Buildkite Agent on Windows.
The old script used a While loop wait until the Buildkite Agent exited.
However, this occasionally failed, probably due to a race condition, so
we ran the finally block even though the agent was still running.

This would cause jobs to fail at random points in time.

By using the native methods Start-Service and WaitForStatus, we can
avoid this problem.