Ignore unexpected fields in Datastore entities
If a new feature requires additional configuration settings, we have to add them to the Datastore entity in order to test everything locally. However, this means that any ongoing deployment fails since it cannot handle the modified entity until the new code is submitted.
This commit solves that problem by ignoring unexpected fields in the Datastore entity.