Always download Buildifier from GitHub.
Previously CI used the Buildifier binary from the Docker container if the user did not request a specific version.
However, since we forgot to update the Docker container, it still contained an ancient version (0.22.0 - for comparison, the most recent release is 4.0.1).

With this commit CI will always download a Buildifier binary from GitHub, with "no version" implying "latest".
Moreover, this commit fixes the download logic to recognize binaries of older releases such as 0.4.3.

Consequently, we no longer ship a Buildifier binary as part of the Docker container.

More progress towards #1080.