No longer crash if no incompatible flags were found
In the new world of LTS and rolling releases we no longer require incompatible flag issues.
Consequently, we should not crash if there are no such flags.

This is a temporary workaround for https://github.com/bazelbuild/bazel/issues/13524.
In the long term we probably want to get rid of the incompatible flag code completely.