Stop removing *all* tasks with custom Bazel versions from downstream pipelines (#1492)
Context: The CI configuration of a downstream pipeline basically
consists of all the CI configurations of its individual projects, but
with a single Bazel version (which depends on the downstream pipeline).

Previously we discarded any tasks that used a specific Bazel version
such as `last_green`. This turned out to be too aggressive for projects
such as rules_apple. As a result, with this commit we only discard tasks
if they would result in duplicate work (i.e. if multiple tasks are
identical except for the Bazel version).

Consequently, we should see more tasks in downstream pipelines, thus
increasing test coverage.

Fixes #1455