Choose Xcode version based on MacOS version
Due to an ongoing infra update we're operating multiple different versions of MacOS, which unfortunately are not all compatible with the same Xcode versions.
Consequently, this commit improves the Xcode version selection logic by determining the Xcode version based on the MacOS version.

This behavior can be avoided by explicitly specifying an Xcode version in the task config.

Eventually we will run the same versions on all machines in the MacOS fleet.

Progress towards #1540