chore: update default branch of rules_nodejs
I'm not sure why these are hardcoded here, when GitHub API provides a way to know the name of the default branch in a repo.
rules_nodejs is preparing a 6.0 which is big breaking change, so we're also taking the opportunity to switch to the standard name for default branch.