chore: update default branch of rules_nodejs (#1581)
I'm not sure why these are hardcoded here, when GitHub API provides a
way to know the name of the default branch in a repo. rules_nodejs is
preparing a 6.0 which is big breaking change, so we're also taking the
opportunity to switch to the standard name for default branch.

NOTE: I've left the `stable` branch on the repo, but it still points to
the old 5.x branch of code, so this change is needed to start testing
the current version.