Add bazel-diff experiment
This experiment reduces test targets to those that have been affected by the commit-under-test, as calculated by bazel-diff.

It is currently guarded by the `EXPERIMENTAL_USE_BAZEL_DIFF` env variable, i.e. it has to be manually enabled when creating a build.

Thanks to our complex CI setup there is no better way of testing this change than "in prod". However, because of the env variable it should not affect our users (famous last words...).