bazel-diff: Use separate workspace for diffbase
Previously we performed all operations in the current workspace directory (including checking out the diffbase commit).
This had some unintended consequences. As a result, this change creates a separate directory for the diffbase commit.
The script simply downloads the .tar.gz archive of the given commit since it should be faster than "git clone".

Progress towards #1643