Decode Test Analytics token before running tests (#1656)
so that bazelci-agent can upload test results to Test Analytics.

Tokens can be acquired from Buildkite's Test Suite and are different for
every Test Suite. We want to supply the token when configuring the
pipelines with Terraform which mean we need to check in the token. We
need to encrypt the token first with command:

```
echo TOKEN | gcloud kms encrypt --project bazel-untrusted --location global --keyring buildkite --key buildkite-untrusted-api-token --ciphertext-file=- --plaintext-file=- | base64
```

The encrypted token can then be checked in with Terraform with env
`ENCRYPTED_BUILDKITE_ANALYTICS_TOKEN`.

This CL allows BazelCI to decrypt the token if is set, before running
the tests.

Working towards Test Analytics integration.