Fix infinite loop in dashboard server (#1731)
The bug was introduced last time I refactored the code using JDK19 and
virtual threads. The infinite loop can happen when the server tries to
sync comments data of one issue from Github and if the number of pages
is greater than 1. It starves the scheduler thread, preventing the
server handles more events and syncing tasks, resulting in the delay of
syncing until a restart.