Add support for Linux ARM64 workers (#2273)
With this change we can create ARM64 GCE images, as well as
multi-platform Ubuntu 20.04 Docker images.

This change does not impact any existing pipelines, but it allows as to
build new images and test them more thoroughly.

Proof of new images in testing org (tests are still failing):
https://buildkite.com/bazel-testing/bazel-bazel-arm64/builds/6#01968447-ec09-493f-9fe9-ac95cb4f2185

Progress towards
https://github.com/bazelbuild/continuous-integration/issues/2272